+++ examples/flask_alchemy/demoapp_factories.py
@@ -1,5 +1,5 @@
 import demoapp
-import factory
+import factory.alchemy
 import factory.fuzzy