+++ docs/changelog.rst
@@ -18,6 +18,7 @@ ChangeLog
     - :issue:`630`: TypeError masked by __repr__ AttributeError when initializing ``Maybe`` with inconsistent phases.
 
 *Removed:*
+
     - Drop support for Python 2 and 3.4. These versions `are not maintained anymore <https://devguide.python.org/devcycle/#end-of-life-branches>`__.
     - Drop support for Django 2.0 and 2.1. These versions `are not maintained anymore <https://www.djangoproject.com/download/#supported-versions>`__.
     - Remove deprecated ``force_flush`` from ``SQLAlchemyModelFactory`` options. Use